requirements.txt
Flask
gunicorn
requests
openai
python-dotenv
redis